refactor

+++ DMS.Application/Patients/Queries/GetPatientById/GetPatientByIdQuery.cs
@@ -1,11 +1,11 @@
 ﻿using DMS.Application.Common.Responses;
-using DMS.Application.DTOs.Patient;
+using DMS.Application.Patients.Models;
 using MediatR;
 
 namespace DMS.Application.Patients.Queries.GetPatientById
 {
    public class GetPatientByIdQuery : IRequest<Response<PatientDto>>
    {
-      public int PatientId { get; init; }
+      public long PatientId { get; init; }
    }
 }
